Move song-request processing into a separate module
